Reset architecture for importer
THe importer module is used for both the modules within importer as well
as srpmproc. A lot of variables, functions, and the like are duplicated
either under similar names or wildly different names that are
unnecessary. This change introduces a dataclass to Import, brings over
majority of expected properties, and a couple of helpers.

A new model was introduced for how data is returned on every import to
almost match what the previous srpmproc does.